Added new features ✨

+++ controllers/goalController.js
@@ -6,7 +6,7 @@ const Goal = require("../models/goalModel");
 // @route   GET /api/goals
 // @access  Private
 const getGoals = asyncHandler(async (req, res) => {
-  const goals = await Goal.find();
+  const goals = await Goal.find({ user: req.user.id }).select("-__v");
   res.status(200).json(goals);
 });
 
@@ -21,6 +21,7 @@ const setGoal = asyncHandler(async (req, res) => {
 
   const goal = await Goal.create({
     title: req.body.title,
+    user: req.user.id,
   });
 
   res.status(200).json(goal);
@@ -36,6 +37,17 @@ const updateGoal = asyncHandler(async (req, res) => {
     throw new Error("Goal not found");
   }
 
+  // Check for user
+  if (!req.user) {
+    res.status(401);
+    throw new Error("User not found");
+  }
+
+  // Make sure the logged in user matches the goal user
+  if (goal.user.toString() !== req.user.id) {
+    res.status(401);
+    throw new Error("User not authorized");
+  }
   const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
     new: true,
   });
@@ -52,13 +64,25 @@ const deleteGoal = asyncHandler(async (req, res) => {
     throw new Error("Goal not found");
   }
 
+  // Check for user
+  if (!req.user) {
+    res.status(401);
+    throw new Error("User not found");
+  }
+
+  // Make sure the logged in user matches the goal user
+  if (goal.user.toString() !== req.user.id) {
+    res.status(401);
+    throw new Error("User not authorized");
+  }
+  
   const deletedGoal = await Goal.findByIdAndRemove(req.params.id, req.body, {
     new: true,
-  })
+  });
   res.status(200).json({
     data: deletedGoal,
     id: req.params.id,
-    message: 'Goal were deleted.'
+    message: "Goal were deleted.",
   });
 });
